Fix bug in handling early exit
Move left-over functions
Add comments
